+++ CatchMate/app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.CatchMate"
         tools:targetApi="31">
         <activity
-            android:name="com.catchmate.presentation.home.MainActivity"
+            android:name="com.catchmate.presentation.view.home.MainActivity"
             android:exported="true"
             tools:replace="android:exported">
             <intent-filter>
