+++ CatchMate/presentation/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
 
     <application>
         <activity
-            android:name=".home.MainActivity"
+            android:name=".view.home.MainActivity"
             android:exported="false" />
     </application>
