+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/view/home/HomeDateFilterBottomSheetFragment.kt
@@ -1,18 +1,18 @@
-package com.catchmate.presentation.home
+package com.catchmate.presentation.view.home
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.catchmate.presentation.R
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-class HomeFragment : Fragment() {
+class HomeDateFilterBottomSheetFragment : BottomSheetDialogFragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_home_date_filter_bottom_sheet, container, false)
     }
 }
