+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/view/home/HomeFilterFragment.kt
@@ -1,4 +1,4 @@
-package com.catchmate.presentation.login
+package com.catchmate.presentation.view.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,12 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import com.catchmate.presentation.R
 
-class LoginFragment : Fragment() {
+class HomeFilterFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login, container, false)
+        return inflater.inflate(R.layout.fragment_home_filter, container, false)
     }
 }
