+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/view/home/HomeTeamFilterBottomSheetFragment.kt
@@ -1,4 +1,4 @@
-package com.catchmate.presentation.home
+package com.catchmate.presentation.view.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,13 @@ import android.view.ViewGroup
 import com.catchmate.presentation.R
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-class HomeDateFilterBottomSheetFragment : BottomSheetDialogFragment() {
+class HomeTeamFilterBottomSheetFragment : BottomSheetDialogFragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_date_filter_bottom_sheet, container, false)
+        return inflater.inflate(R.layout.fragment_home_team_filter_bottom_sheet, container, false)
     }
+    // 체크박스마다 setOnCheckedChangeListener 달아서 이미지뷰 리소스를 변경하기
 }
