+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/view/login/LoginFragment.kt
@@ -1,0 +1,18 @@
+package com.catchmate.presentation.view.login
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.catchmate.presentation.R
+
+class LoginFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_login, container, false)
+    }
+}
