+++ CatchMate/presentation/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#F7F9FA"
-    tools:context=".home.HomeFragment" >
+    tools:context=".view.home.HomeFragment" >
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline_home_left"
