+++ CatchMate/presentation/src/main/res/layout/fragment_home_filter.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center"
-    tools:context=".home.HomeFilterFragment">
+    tools:context=".view.home.HomeFilterFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline_home_filter_left"
