+++ CatchMate/presentation/src/main/res/layout/fragment_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.LoginFragment" >
+    tools:context=".view.login.LoginFragment" >
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline_login_left"
